requirements.txt
aenum==2.2.4
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
dbf==0.99.0
decorator==4.4.1
defusedxml==0.6.0
Django==1.11.26
django-contrib-comments==1.9.2
django-cors-headers==3.2.0
django-extensions==2.2.5
djangorestframework==3.11.0
docopt==0.6.2
entrypoints==0.3
filebrowser-safe==0.5.0
future==0.18.2
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
graphviz==0.13.2
grappelli-safe==0.5.2
httplib2==0.17.1
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
# lxml==4.4.2
MarkupSafe==1.1.1
Mezzanine==4.3.1
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
num2words==0.5.10
numpy==1.20.3
oauthlib==3.1.0
pandas==1.2.4
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.2.0
plivo==4.3.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.3
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
Pygments==2.5.2
PyPDF2==1.26.0
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
reportlab==3.5.67
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
Send2Trash==1.5.0
six==1.13.0
soupsieve==1.9.5
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
XlsxWriter==1.2.8
zipp==0.6.0
